split lib and cli
