+++ Cargo.toml
@@ -1,22 +1,5 @@
-[package]
-name = "sulu"
-version = "0.1.2"
-authors = ["Tom Watson <tom.watson@kinesis.org>"]
-edition = "2018"
-license = "MIT"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = { version = "^1.0", features = ["derive"] }
-serde_json = "^1.0"
-osmpbfreader = "^0.14"
-geo-types = "^0.6"
-geo = "^0.14"
-gdal = { version = "^0.7.2", optional=true }
-geojson = { version = "^0.19", features = ["geo-types"] }
-smartstring = "^0.2.3"
-clap = "^2.33"
-
-[features]
-default = ["gdal"]
+[workspace]
+members = [
+    "sulu-cli",
+    "sulu-lib"
+]
