+++ sulu-cli/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "sulu"
+version = "0.1.2"
+authors = ["Tom Watson <tom.watson@kinesis.org>"]
+edition = "2018"
+license = "MIT"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+gdal = { version = "^0.7.2", optional=true }
+geo-types = "^0.6"
+geojson = { version = "^0.19", features = ["geo-types"], optional=true }
+clap = "^2.33"
+serde_json = "^1.0"
+sulu-lib = { path = "../sulu-lib/" }
+
+[features]
+default = ["formats-gdal", "formats-geojson"]
+formats-gdal = ["gdal", "sulu-lib/formats-gdal"]
+formats-geojson = ["geojson", "sulu-lib/formats-geojson"]
