+++ sulu-lib/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sulu-lib"
+version = "0.1.2"
+authors = ["Tom Watson <tom.watson@kinesis.org>"]
+edition = "2018"
+license = "MIT"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "^1.0", features = ["derive"] }
+serde_json = "^1.0"
+osmpbfreader = "^0.14"
+geo-types = "^0.6"
+geo = "^0.14"
+smartstring = "^0.2.3"
+geojson = { version = "^0.19", features = ["geo-types"], optional=true }
+gdal = { version = "^0.7.2", optional=true }
+
+[features]
+formats-gdal = ["gdal"]
+formats-geojson = ["geojson"]
