+++ sulu-lib/src/edge_list.rs
@@ -13,6 +13,11 @@ use crate::{
     osm_cache::OSMCache,
     error::Error,
 };
+use serde_json::{json, Value, Map};
+#[cfg(feature="formats-geojson")]
+use geojson;
+#[cfg(feature="formats-gdal")]
+use gdal;
 
 #[derive(Debug)]
 pub struct Edge<T: CoordinateType> {
@@ -156,4 +161,53 @@ fn edges_from_osm_id(
 
     edges.shrink_to_fit();
     Ok(Some(edges))
+}
+
+
+#[cfg(feature="formats-geojson")]
+impl From<Edge<f64>> for geojson::Feature {
+    fn from(edge: Edge<f64>) -> geojson::Feature {
+        let mut props: Map<String, Value> = Map::new();
+        props.insert("way_osmid".to_string(), json!(edge.way_osmid));
+        props.insert("start_node_id".to_string(), json!(edge.start_node_id));
+        props.insert("end_node_id".to_string(), json!(edge.end_node_id));
+        props.insert("graph_config_option".to_string(), json!(edge.graph_config_option.name));
+        props.insert("length_m".to_string(), json!(edge.length_m));
+        let geom = geojson::Geometry {
+            bbox: None,
+            foreign_members: None,
+            value: (&edge.geometry).into()
+        };
+        geojson::Feature {
+            bbox: None,
+            geometry: Some(geom),
+            id: None,
+            properties: Some(props),
+            foreign_members: None
+        }
+    }
+}
+
+#[cfg(feature="formats-geojson")]
+impl From<EdgeList<f64>> for geojson::FeatureCollection {
+    fn from(el: EdgeList<f64>) -> geojson::FeatureCollection {
+        geojson::FeatureCollection {
+            bbox: None,
+            foreign_members: None,
+            features: el.edges.into_iter().map(|e| e.into()).collect()
+        }
+    }
+}
+
+#[cfg(feature="formats-gdal")]
+pub fn linestring_to_gdal(linestring: &LineString<f64>) -> Result<gdal::vector::Geometry, Error> {
+    // gdal version 0.6.0 relies on geo-types 0.4.0 (not 0.6.0), but I don't want to change
+    // versions of geo-types, so we can just hack this in here, and use the ToGdal
+    // trait when gdal catches up
+    let mut geom = gdal::vector::Geometry::empty(gdal::vector::OGRwkbGeometryType::wkbLineString)
+        .map_err(Error::GdalError)?;
+    for (i, pt) in linestring.points_iter().enumerate() {
+        geom.set_point_2d(i, (pt.x(), pt.y()));
+    }
+    Ok(geom)
 }
