+++ sulu-lib/src/error.rs
@@ -8,7 +8,7 @@ pub enum Error {
     NotANode(osmpbfreader::objects::OsmObj),
     NotAWayId(osmpbfreader::objects::OsmId),
     ConversionError(String),
-    #[cfg(feature="gdal")]
+    #[cfg(feature="formats-gdal")]
     GdalError(gdal::errors::GdalError),
     NodeCountError,
     MakeGraphError
@@ -29,7 +29,7 @@ impl std::fmt::Display for Error {
                 write!(f, "OsmId is not a way id: {:?}", o),
             Error::ConversionError(s) =>
                 write!(f, "ConversionError: {}", s),
-            #[cfg(feature="gdal")]
+            #[cfg(feature="formats-gdal")]
             Error::GdalError(e) =>
                 write!(f, "GdalError: {}", e),
             Error::NodeCountError =>
