+++ sulu-lib/src/graph_config.rs
